+++ firmware/config/boards/hellen/cypress/config/controllers/algo/engine_configuration_generated_structures.h
@@ -1,4 +1,4 @@
-// this section was generated automatically by rusEfi tool ConfigDefinition.jar based on hellen_cypress_gen_config.bat integration/rusefi_config.txt Tue Nov 10 17:38:43 UTC 2020
+// this section was generated automatically by rusEfi tool ConfigDefinition.jar based on hellen_cypress_gen_config.bat integration/rusefi_config.txt Tue Nov 17 00:53:37 UTC 2020
 // by class com.rusefi.output.CHeaderConsumer
 // begin
 #pragma once
@@ -604,8 +604,8 @@ struct idle_hardware_s {
 
 typedef struct idle_hardware_s idle_hardware_s;
 
-// start of etb_io
-struct etb_io {
+// start of dc_io
+struct dc_io {
 	/**
 	 * offset 0
 	 */
@@ -625,7 +625,7 @@ struct etb_io {
 	/** total size 4*/
 };
 
-typedef struct etb_io etb_io;
+typedef struct dc_io dc_io;
 
 // start of engine_configuration_s
 struct engine_configuration_s {
@@ -751,14 +751,14 @@ struct engine_configuration_s {
 	offset 76 bit 31 */
 	bool unusedBit_34_31 : 1;
 	/**
-	 * Closed throttle. todo: extract these two fields into a structure
+	 * Closed throttle, 1 volt = 200 units.
 	 * See also tps1_1AdcChannel
 	 * set tps_min X
 	 * offset 80
 	 */
 	int16_t tpsMin;
 	/**
-	 * Full throttle. tpsMax value as 10 bit ADC value. Not Voltage!
+	 * Full throttle.
 	 * See also tps1_1AdcChannel
 	 * set tps_max X
 	 * offset 82
@@ -1338,6 +1338,7 @@ struct engine_configuration_s {
 	/**
 	 * Voltage when the wastegate is fully open.
 	 * You probably don't have one of these!
+	 * 1 volt = 1000 units
 	 * offset 718
 	 */
 	uint16_t wastegatePositionMax;
@@ -1350,6 +1351,7 @@ struct engine_configuration_s {
 	/**
 	 * Voltage when the idle valve is open.
 	 * You probably don't have one of these!
+	 * 1 volt = 1000 units
 	 * offset 722
 	 */
 	uint16_t idlePositionMax;
@@ -1858,7 +1860,7 @@ struct engine_configuration_s {
 	/**
 	 * offset 980
 	 */
-	etb_io etbIo[ETB_COUNT];
+	dc_io etbIo[ETB_COUNT];
 	/**
 	 * Wastegate control Solenoid
 	 * offset 988
@@ -1945,7 +1947,7 @@ struct engine_configuration_s {
 	/**
 	 * offset 1088
 	 */
-	etb_io etbIo2[ETB_COUNT];
+	dc_io stepperDcIo[DC_PER_STEPPER];
 	/**
 	 * For example, BMW, GM or Chevrolet
 	 * REQUIRED for rusEFI Online
@@ -2788,7 +2790,7 @@ struct engine_configuration_s {
 	/**
 	 * offset 2130
 	 */
-	brain_pin_e hpfpValvePin;
+	output_pin_e hpfpValvePin;
 	/**
 	 * offset 2131
 	 */
@@ -2924,11 +2926,17 @@ struct engine_configuration_s {
 	/**
 	 * offset 2418
 	 */
-	uint8_t unusedSomethingWasHere[2];
+	injector_compensation_mode_e injectorCompensationMode;
 	/**
+	 * offset 2419
+	 */
+	uint8_t unused2419;
+	/**
+	 * This is the pressure at which your injector flow is known.
+	 * For example if your injectors flow 400cc/min at 3.5 bar, enter 350kpa here.
 	 * offset 2420
 	 */
-	float unused244_1;
+	float fuelReferencePressure;
 	/**
 	 * offset 2424
 	 */
@@ -3761,4 +3769,4 @@ struct persistent_config_s {
 typedef struct persistent_config_s persistent_config_s;
 
 // end
-// this section was generated automatically by rusEfi tool ConfigDefinition.jar based on hellen_cypress_gen_config.bat integration/rusefi_config.txt Tue Nov 10 17:38:43 UTC 2020
+// this section was generated automatically by rusEfi tool ConfigDefinition.jar based on hellen_cypress_gen_config.bat integration/rusefi_config.txt Tue Nov 17 00:53:37 UTC 2020