+++ firmware/config/engines/mazda_miata_1_6.cpp
@@ -421,7 +421,6 @@ void setMiataNA6_MAP_MRE(DECLARE_CONFIG_PARAMETER_SIGNATURE) {
 	engineConfiguration->idleTimingPid.dFactor = 0.0;
 	engineConfiguration->idleTimingPid.minValue = -13;
 	engineConfiguration->idleTimingPid.maxValue = 13;
-	engineConfiguration->idleTimingPid.periodMs = 8;
 	engineConfiguration->idleTimingPidWorkZone = 150;
 	engineConfiguration->idlePidFalloffDeltaRpm = 50;
 	engineConfiguration->idleTimingPidDeadZone = 10;