+++ firmware/config/engines/vw.cpp
@@ -16,6 +16,7 @@
 EXTERN_CONFIG;
 
 // VW_ABA
+// set engine_type 32
 void setVwAba(DECLARE_CONFIG_PARAMETER_SIGNATURE) {
 	setFrankensoConfiguration(PASS_CONFIG_PARAMETER_SIGNATURE);
 
@@ -51,6 +52,9 @@ void setVwAba(DECLARE_CONFIG_PARAMETER_SIGNATURE) {
 	engineConfiguration->ignitionPins[2] = GPIO_UNASSIGNED;
 	engineConfiguration->ignitionPins[3] = GPIO_UNASSIGNED;
 	engineConfiguration->ignitionPinMode = OM_DEFAULT;
+
+	engineConfiguration->idlePositionSensor = EFI_ADC_3; // PA3
+	engineConfiguration->wastegatePositionSensor = EFI_ADC_4; // PA4
 
 	float mapRange = 110;