+++ firmware/controllers/actuators/idle_hardware.cpp
@@ -139,8 +139,8 @@ void initIdleHardware(Logging* sharedLogger DECLARE_ENGINE_PARAMETER_SUFFIX) {
 		StepperHw* hw;
 
 		if (CONFIG(useHbridges)) {
-			auto motorA = initDcMotor(2, /*useTwoWires*/ true PASS_ENGINE_PARAMETER_SUFFIX);
-			auto motorB = initDcMotor(3, /*useTwoWires*/ true PASS_ENGINE_PARAMETER_SUFFIX);
+			auto motorA = initDcMotor(engineConfiguration->stepperDcIo[0], 2, /*useTwoWires*/ true PASS_ENGINE_PARAMETER_SUFFIX);
+			auto motorB = initDcMotor(engineConfiguration->stepperDcIo[1], 3, /*useTwoWires*/ true PASS_ENGINE_PARAMETER_SUFFIX);
 
 			if (motorA && motorB) {
 				iacHbridgeHw.initialize(
@@ -174,25 +174,22 @@ void initIdleHardware(Logging* sharedLogger DECLARE_ENGINE_PARAMETER_SUFFIX) {
 		 */
 		// todo: even for double-solenoid mode we can probably use same single SimplePWM
 		// todo: open question why do we pass 'OutputPin' into 'startSimplePwmExt' if we have custom applyIdleSolenoidPinState listener anyway?
-		if (!CONFIG(isDoubleSolenoidIdle)) {
-			startSimplePwm(&idleSolenoidOpen, "Idle Valve",
-					&engine->executor,
-					&enginePins.idleSolenoidPin,
-					CONFIG(idle).solenoidFrequency, PERCENT_TO_DUTY(CONFIG(manIdlePosition)),
-					(pwm_gen_callback*)applyIdleSolenoidPinState);
-		} else {
+		if (CONFIG(idle).solenoidPin != GPIO_UNASSIGNED) {
 			startSimplePwm(&idleSolenoidOpen, "Idle Valve Open",
-					&engine->executor,
-					&enginePins.idleSolenoidPin,
-					CONFIG(idle).solenoidFrequency, PERCENT_TO_DUTY(CONFIG(manIdlePosition)),
-					(pwm_gen_callback*)applyIdleSolenoidPinState);
-
-			startSimplePwm(&idleSolenoidClose, "Idle Valve Close",
-					&engine->executor,
-					&enginePins.secondIdleSolenoidPin,
-					CONFIG(idle).solenoidFrequency, PERCENT_TO_DUTY(CONFIG(manIdlePosition)),
-					(pwm_gen_callback*)applyIdleSolenoidPinState);
+				&engine->executor,
+				&enginePins.idleSolenoidPin,
+				CONFIG(idle).solenoidFrequency, PERCENT_TO_DUTY(CONFIG(manIdlePosition)),
+				(pwm_gen_callback*)applyIdleSolenoidPinState);
 		}
+
+		if (CONFIG(isDoubleSolenoidIdle) && CONFIG(secondSolenoidPin) != GPIO_UNASSIGNED) {
+			startSimplePwm(&idleSolenoidClose, "Idle Valve Close",
+				&engine->executor,
+				&enginePins.secondIdleSolenoidPin,
+				CONFIG(idle).solenoidFrequency, PERCENT_TO_DUTY(CONFIG(manIdlePosition)),
+				(pwm_gen_callback*)applyIdleSolenoidPinState);
+		}
+
 		idlePositionSensitivityThreshold = 0.0f;
 	}
 }